backend/requirements.txt
# Flask core
Flask==3.0.0
Werkzeug==3.0.1

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg[binary]==3.2.13

# Authentication
Flask-JWT-Extended==4.6.0

# CORS
Flask-CORS==4.0.0

# Serialization/Validation
marshmallow==3.20.1
flask-marshmallow==0.15.0

# Environment variables
python-dotenv==1.0.0

